# Conflicts:
#	scripts/cmake/vcpkg_configure_cmake.cmake
#	scripts/cmake/vcpkg_find_acquire_program.cmake
#	toolsrc/src/vcpkg_System.cpp

+++ scripts/internalCI.ps1
@@ -5,7 +5,10 @@ New-Item -type directory downloads -errorAction SilentlyContinue | Out-Null
 if (-not $?) { throw $? }
 
 # Clear out any intermediate files from the previous build
-Get-ChildItem buildtrees/*/* | ? { $_.Name -ne "src" -and $_.Extension -ne ".log"} | Remove-Item -Recurse -Force
+if (Test-Path buildtrees)
+{
+    Get-ChildItem buildtrees/*/* | ? { $_.Name -ne "src" -and $_.Extension -ne ".log"} | Remove-Item -Recurse -Force
+}
 
 # Purge any outdated packages
 ./vcpkg remove --outdated --recurse
@@ -14,7 +17,7 @@ if (-not $?) { throw $? }
 ./vcpkg.exe install azure-storage-cpp cpprestsdk:x64-windows-static cpprestsdk:x86-uwp
 if (-not $?) { throw $? }
 
-./vcpkg.exe install bond chakracore cryptopp zlib expat sdl2 curl sqlite3 libuv protobuf:x64-windows sfml opencv:x64-windows uwebsockets uwebsockets:x64-windows-static
+./vcpkg.exe install bond cryptopp zlib expat sdl2 curl sqlite3 libuv protobuf:x64-windows sfml opencv:x64-windows uwebsockets uwebsockets:x64-windows-static
 if (-not $?) { throw $? }
 
 ./vcpkg.exe install opencv:x86-uwp boost:x86-uwp
